+++ f-list.h
@@ -79,6 +79,7 @@ typedef struct FListFriends_ FListFriends;
 #define FLIST_PORT_SECURE       9799
 #define GLOBAL_NAME             "#FList"
 #define FLIST_DEBUG             "flist"
+#define FLIST_PIDGIN_CHANNEL    "ADH-1509b63f1b1a98c79ab9"
 
 #define CONVO_SHOW_CHAT         "CONVO_SHOW_CHAT"
 #define CONVO_SHOW_ADS          "CONVO_SHOW_ADS"